+++ internal/handler/company_handler/create_company_handler.go
@@ -1,0 +1,50 @@
+package company_handler
+
+import (
+	"blueberry_homework/dto/request"
+	"blueberry_homework/dto/response"
+	"encoding/json"
+	"net/http"
+)
+
+func (h *CompanyHandler) CreateCompany(w http.ResponseWriter, r *http.Request) {
+	var req request.CreateCompanyRequest
+
+	// null check validation
+	err := json.NewDecoder(r.Body).Decode(&req)
+	if err != nil || req.CompanyName == "" || req.UserID == "" || req.CompanyAddress == "" {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusBadRequest)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   "Invalid request format",
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+			return
+		}
+		return
+	}
+
+	err = h.CompanyUsecase.CreateCompany(req)
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusBadRequest)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   err.Error(),
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+			return
+		}
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
+	if err := json.NewEncoder(w).Encode(response.SuccessResponse{
+		Message: "company created successfully",
+	}); err != nil {
+		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		return
+	}
+}
