+++ internal/handler/user_handler/get_user_handler.go
@@ -1,0 +1,78 @@
+package user_handler
+
+import (
+	"blueberry_homework/dto/response"
+	"blueberry_homework/middleware"
+	"encoding/json"
+	"net/http"
+
+	"github.com/golang-jwt/jwt/v5"
+)
+
+func (h *UserHandler) GetUser(w http.ResponseWriter, r *http.Request) {
+	// 미들웨어에서 저장한 claims를 context에서 가져오기
+	claimsValue := r.Context().Value(middleware.ClaimsContextKey)
+	if claimsValue == nil {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusUnauthorized)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   "JWT claims not found in context",
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		}
+		return
+	}
+
+	claims, ok := claimsValue.(jwt.MapClaims)
+	if !ok {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusInternalServerError)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   "Failed to parse JWT claims from context",
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		}
+		return
+	}
+
+	// sub claim에서 user ID 추출
+	userID, ok := claims["sub"].(string)
+	if !ok {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusBadRequest)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   "Invalid user ID in token",
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		}
+		return
+	}
+
+	// Get User Data
+	user, err := h.usecase.GetUser(userID)
+	if err != nil {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusInternalServerError)
+		if err := json.NewEncoder(w).Encode(response.ErrorResponse{
+			Message: "error",
+			Error:   "Failed to get user data",
+		}); err != nil {
+			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		}
+		return
+	}
+
+	// RETURN USER DATA
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	if err := json.NewEncoder(w).Encode(response.GetUserResponse{
+		Message: "success",
+		Data:    user,
+	}); err != nil {
+		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
+		return
+	}
+}
