+++ route/user_route.go
@@ -1,19 +1,24 @@
 package route
 
 import (
-	"blueberry_homework/internal/handler"
+	"blueberry_homework/internal/handler/user_handler"
+	"blueberry_homework/middleware"
 
 	"github.com/go-chi/chi/v5"
 )
 
-func UserRouter(h *handler.UserHandler) chi.Router {
+func UserRouter(h *user_handler.UserHandler) chi.Router {
 	r := chi.NewRouter()
 
-	// user
+	// 토큰 인증 필요 없는 라우트
 	r.Post("/sign-up", h.SignUp)
 	r.Post("/log-in", h.Login)
-	r.Get("/get-user", h.GetUser)
-	r.Put("/change-user", h.ChangeUser)
 
+	// 토큰 인증이 필요한 라우트 (protected route)
+	r.Group(func(pr chi.Router) {
+		pr.Use(middleware.VerifyToken)
+		pr.Get("/get-user", h.GetUser)
+		pr.Put("/change-user", h.ChangeUser)
+	})
 	return r
 }
